tidy(ui): move viewerMode to uiSlice
Don't need a whole slice for this.
