+++ invokeai/frontend/web/src/app/store/store.ts
@@ -26,7 +26,6 @@ import { sdxlPersistConfig, sdxlSlice } from 'features/sdxl/store/sdxlSlice';
 import { configSlice } from 'features/system/store/configSlice';
 import { systemPersistConfig, systemSlice } from 'features/system/store/systemSlice';
 import { uiPersistConfig, uiSlice } from 'features/ui/store/uiSlice';
-import { viewerPersistConfig, viewerSlice } from 'features/viewer/store/viewerSlice';
 import { diff } from 'jsondiffpatch';
 import { defaultsDeep, keys, omit, pick } from 'lodash-es';
 import dynamicMiddlewares from 'redux-dynamic-middlewares';
@@ -63,7 +62,6 @@ const allReducers = {
   [queueSlice.name]: queueSlice.reducer,
   [workflowSlice.name]: workflowSlice.reducer,
   [hrfSlice.name]: hrfSlice.reducer,
-  [viewerSlice.name]: viewerSlice.reducer,
   [progressSlice.name]: progressSlice.reducer,
   [api.reducerPath]: api.reducer,
 };
@@ -109,7 +107,6 @@ const persistConfigs: { [key in keyof typeof allReducers]?: PersistConfig } = {
   [loraPersistConfig.name]: loraPersistConfig,
   [modelManagerPersistConfig.name]: modelManagerPersistConfig,
   [hrfPersistConfig.name]: hrfPersistConfig,
-  [viewerPersistConfig.name]: viewerPersistConfig,
   [progressPersistConfig.name]: progressPersistConfig,
 };
